+++ CHANGELOG.md
@@ -17,7 +17,8 @@
 - Generate a new random id for archived formulas
 - Documentation links update
 - Session monitor now logs to a file if defined
-
+- Improved CSV formula import
+ 
 ### Version 12.8
 - Added links for PV apps
 - Fix ingredient import compounds failing to import