+++ releasenotes.md
@@ -9,4 +9,5 @@ Whats New in v12.9
 - Removed time from created and updated fields
 - Removed created and updated timestamps from ingredient suppliers, bottles, customers and formula documents
 - Removed updated timestamp from html templates
+- Fixed CSV formula import
 - This release may include additional changes; for full details, please refer to the CHANGELOG